requirements.txt
addict==2.4.0
Box2D==2.3.10
Box2D-kengz==2.3.3
certifi==2022.12.7
charset-normalizer==3.0.1
cloudpickle==2.2.1
cycler==0.11.0
dictor==0.1.11
ephem==4.1.4
fonttools==4.38.0
gym==0.26.2
gym-notices==0.0.8
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
kiwisolver==1.4.4
matplotlib==3.5.3
networkx==2.6.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
protobuf==3.20.3
py-trees==0.8.3
pydot==1.4.2
pyparsing==3.0.9
PySnooper==1.1.1
python-dateutil==2.8.2
PyWavelets==1.3.0
requests==2.28.2
scikit-image==0.19.3
scipy==1.7.3
shapely==2.0.1
six==1.16.0
tensorboardX==2.6
tifffile==2021.11.2
torch==1.13.1
torchaudio==0.13.1
TorchSnooper==0.8
torchvision==0.14.1
typing_extensions==4.5.0
urllib3==1.26.14
yapf==0.32.0
zipp==3.15.0
